requirements.txt
google-auth>=1.11.0
google-cloud-pubsub>=1.1.0
google-cloud-core>=1.2.0
mock>=3.0.5
oauth2client>=4.1.3
requests>=2.22.0
